#125 EA-1605 Refactored atomic update methods.
